feat(timer): mover timer to specific component
